+++ R/grid_code_paths.R
@@ -6,6 +6,10 @@ tune_nothing_with_formula <- function(resamples, grid, workflow, metrics, contro
   resample_with_formula(resamples, workflow, metrics, control)
 }
 
+tune_nothing_with_variables <- function(resamples, grid, workflow, metrics, control)  {
+  resample_with_variables(resamples, workflow, metrics, control)
+}
+
 # ------------------------------------------------------------------------------
 
 iter_rec_and_mod <- function(rs_iter, resamples, grid, workflow, metrics, control) {
@@ -353,7 +357,7 @@ iter_mod_with_recipe <- function(rs_iter, resamples, grid, workflow, metrics, co
     notes = .notes
   )
 
-  # check for recipe failure
+  # check for failure
   if (is_failure(workflow)) {
     out <- list(
       .metrics = metric_est,
@@ -392,7 +396,7 @@ iter_mod_with_recipe <- function(rs_iter, resamples, grid, workflow, metrics, co
     )
 
     workflow <- catch_and_log_fit(
-      train_model(workflow, mod_grid_vals[mod_iter,], control_workflow),
+      train_model(workflow, param_val, control_workflow),
       control,
       split,
       mod_msg,
@@ -407,21 +411,22 @@ iter_mod_with_recipe <- function(rs_iter, resamples, grid, workflow, metrics, co
     # Extract names from the mold
     all_outcome_names <- append_outcome_names(all_outcome_names, workflow)
 
-    extracted <-
-      append_extracts(
-        extracted,
-        workflow,
-        mod_grid_vals[mod_iter, ],
-        split,
-        control,
-        mod_id
-      )
+    extracted <- append_extracts(
+      extracted,
+      workflow,
+      param_val,
+      split,
+      control,
+      mod_id
+    )
+
+    pred_msg <- paste(mod_msg, "(predictions)")
 
     tmp_pred <- catch_and_log(
-      predict_model(split, workflow, mod_grid_vals[mod_iter,], metrics),
+      predict_model(split, workflow, param_val, metrics),
       control,
       split,
-      paste(mod_msg, "(predictions)"),
+      pred_msg,
       bad_only = TRUE,
       notes = .notes
     )
@@ -493,7 +498,7 @@ iter_mod_with_formula <- function(rs_iter, resamples, grid, workflow, metrics, c
     notes = .notes
   )
 
-  # check for formula failure
+  # check for failure
   if (is_failure(workflow)) {
     out <- list(
       .metrics = metric_est,
@@ -547,24 +552,25 @@ iter_mod_with_formula <- function(rs_iter, resamples, grid, workflow, metrics, c
     # Extract names from the mold
     all_outcome_names <- append_outcome_names(all_outcome_names, workflow)
 
-    extracted <-
-      append_extracts(extracted,
-                      workflow,
-                      param_val,
-                      split,
-                      control,
-                      mod_id)
+    extracted <- append_extracts(
+      extracted,
+      workflow,
+      param_val,
+      split,
+      control,
+      mod_id
+    )
 
     pred_msg <- paste(mod_msg, "(predictions)")
 
-    tmp_pred <-
-      catch_and_log(
-        predict_model(split, workflow, param_val, metrics),
-        control,
-        split,
-        mod_msg,
-        notes = .notes
-      )
+    tmp_pred <- catch_and_log(
+      predict_model(split, workflow, param_val, metrics),
+      control,
+      split,
+      pred_msg,
+      bad_only = TRUE,
+      notes = .notes
+    )
 
     # check for prediction level failure
     if (is_failure(tmp_pred)) {
@@ -585,7 +591,150 @@ iter_mod_with_formula <- function(rs_iter, resamples, grid, workflow, metrics, c
   )
 }
 
-# ----------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
+
+tune_mod_with_variables <- function(resamples, grid, workflow, metrics, control) {
+  B <- nrow(resamples)
+
+  `%op%` <- get_operator(control$allow_par, workflow)
+
+  safely_iter_mod_with_variables <- super_safely_iterate(iter_mod_with_variables)
+
+  load_pkgs <- c(control$pkgs, required_pkgs(workflow))
+
+  results <- foreach::foreach(rs_iter = 1:B,
+                              .packages = load_pkgs,
+                              .errorhandling = "pass") %op% {
+    safely_iter_mod_with_variables(rs_iter, resamples, grid, workflow, metrics, control)
+  }
+
+  resamples <- pull_metrics(resamples, results, control)
+  resamples <- pull_notes(resamples, results, control)
+  resamples <- pull_extracts(resamples, results, control)
+  resamples <- pull_predictions(resamples, results, control)
+  resamples <- pull_all_outcome_names(resamples, results)
+
+  resamples
+}
+
+iter_mod_with_variables <- function(rs_iter, resamples, grid, workflow, metrics, control) {
+  load_pkgs(workflow)
+  load_namespace(control$pkgs)
+
+  control_parsnip <- parsnip::control_parsnip(verbosity = 0, catch = TRUE)
+  control_workflow <- control_workflow(control_parsnip = control_parsnip)
+
+  split <- resamples$splits[[rs_iter]]
+  metric_est <- NULL
+  extracted <- NULL
+  pred_vals <- NULL
+  all_outcome_names <- list()
+  .notes <- NULL
+
+  # ----------------------------------------------------------------------------
+
+  workflow <- catch_and_log(
+    train_variables(split, workflow),
+    control,
+    split,
+    "variables",
+    notes = .notes
+  )
+
+  # check for failure
+  if (is_failure(workflow)) {
+    out <- list(
+      .metrics = metric_est,
+      .extracts = extracted,
+      .predictions = pred_vals,
+      .all_outcome_names = all_outcome_names,
+      .notes = .notes
+    )
+
+    return(out)
+  }
+
+  # ----------------------------------------------------------------------------
+
+  # Determine the _minimal_ number of models to fit in order to get
+  # predictions on all models.
+  mod_grid_vals <- workflows::pull_workflow_spec(workflow) %>% min_grid(grid)
+
+  num_mod <- nrow(mod_grid_vals)
+  num_submodels <- mod_grid_vals %>%
+    unnest(.submodels, keep_empty = TRUE) %>%
+    pull(.submodels) %>%
+    map_int(length)
+
+  original_workflow <- workflow
+
+  for (mod_iter in 1:num_mod) {
+    workflow <- original_workflow
+
+    param_val <- mod_grid_vals[mod_iter, ]
+    submodel_id <- mod_iter + sum(vec_slice(num_submodels, 1:mod_iter))
+    mod_msg <- paste0("model ", format(1:num_mod)[mod_iter], "/", num_mod)
+    mod_id <- vec_slice(
+      recipes::names0(nrow(grid), "Model"),
+      (submodel_id - num_submodels[mod_iter]):submodel_id
+    )
+
+    workflow <- catch_and_log_fit(
+      train_model(workflow, param_val, control = control_workflow),
+      control,
+      split,
+      mod_msg,
+      notes = .notes
+    )
+
+    # check for parsnip level and model level failure
+    if (is_failure(workflow) || is_failure(workflow$fit$fit$fit)) {
+      next
+    }
+
+    # Extract names from the mold
+    all_outcome_names <- append_outcome_names(all_outcome_names, workflow)
+
+    extracted <- append_extracts(
+      extracted,
+      workflow,
+      param_val,
+      split,
+      control,
+      mod_id
+    )
+
+    pred_msg <- paste(mod_msg, "(predictions)")
+
+    tmp_pred <- catch_and_log(
+      predict_model(split, workflow, param_val, metrics),
+      control,
+      split,
+      pred_msg,
+      bad_only = TRUE,
+      notes = .notes
+    )
+
+    # check for prediction level failure
+    if (is_failure(tmp_pred)) {
+      next
+    }
+
+    metric_est  <- append_metrics(metric_est, tmp_pred, workflow, metrics, split, mod_id)
+    config_id <- extract_config(workflow, metric_est)
+    pred_vals <- append_predictions(pred_vals, tmp_pred, split, control, config_id)
+  } # end model loop
+
+  list(
+    .metrics = metric_est,
+    .extracts = extracted,
+    .predictions = pred_vals,
+    .all_outcome_names = all_outcome_names,
+    .notes = .notes
+  )
+}
+
+# ------------------------------------------------------------------------------
 
 super_safely_iterate <- function(fn) {
   purrr::partial(.f = super_safely_iterate_impl, fn = fn)